New contract address
Changed contract address to latest version. The latest version of the smart contract supports starting an auction for an item. Auctions are not fully functional though. This is just to test that starting the auction works, there is no functionality to end the auction or to bid for the item.