Added startAuction function
User's can now start auctions. This functionality is not fully implemented. If you start an auction, you will not be able to end it or set an end time. The item will be buyable for the "starting ask" that you enter at item creation. The auction functionality will be more fully implemented as each addition is tested and approved.