Updated Marketplace contract
Added fields to itemAuctionStarted event to track the start time of the auction and the end time. The end time is currently set at 60,000 milliseconds (1 minute) after the start time, just for testing purposes. Also added a function for users to bid on items. Not currently functional, just for testing.